+++ megalinter/reporters/GithubCommentReporter.py
@@ -62,7 +62,7 @@ class GithubCommentReporter(Reporter):
                     status = (
                         "✅"
                         if linter.status == "success" and linter.return_code == 0
-                        else ":orange_circle:"
+                        else ":warning:"
                         if linter.status != "success" and linter.return_code == 0
                         else "❌"
                     )
@@ -109,7 +109,13 @@ class GithubCommentReporter(Reporter):
                 headers=table_header, value_matrix=table_data_raw
             )
             table_content = str(writer) + os.linesep if len(table_data_raw) > 1 else ""
-            status = "✅" if self.master.return_code == 0 else "❌"
+            status = (
+                "✅"
+                if self.master.return_code == 0 and self.master.status == "success"
+                else ":warning:"
+                if self.master.status == "warning"
+                else "❌"
+            )
             status_with_href = (
                 status
                 + " "